refactor: adapter로 명칭 수정
