+++ src/main/java/kitchenpos/deliveryorder/adapter/in/web/DefaultKitchenridersClient.java
@@ -1,0 +1,14 @@
+package kitchenpos.deliveryorder.adapter.in.web;
+
+import kitchenpos.deliveryorder.adapter.out.api.KitchenridersClient;
+import org.springframework.stereotype.Component;
+
+import java.math.BigDecimal;
+import java.util.UUID;
+
+@Component
+public class DefaultKitchenridersClient implements KitchenridersClient {
+    @Override
+    public void requestDelivery(final UUID orderId, final BigDecimal amount, final String deliveryAddress) {
+    }
+}
